Register plugin with soong
Make Soong call a plugin initialization function so that we can
register our modules.

Use android.OncePer to just load the config once. getConfig() needs to
be updated so that it can cope with being called with either Blueprint
contexts or Android contexts.

This commit depends on a Soong commit implementing
android.RegisterPlugin.

When loading Bob's configuration, use the Android configuration to
identify the file to load. Look in both build output and source
directories for the configuration. Generally we prefer having the
configuration in the build output directory. Allowing the
configuration to be found in the source directory means it can be
committed, and (with a few more changes) a build can be done without a
bootstrap call for Bob.

Change-Id: I12dcbcdcb02c04400c3a6b7c9a6250f6a41e6fa5
Signed-off-by: David Kilroy <david.kilroy@arm.com>